feat: Unblock USB ports on Windows
This commit adds a PowerShell script, `unblockusb.ps1`, that can be used to unblock USB ports on Windows. The script checks the current value of the USBSTOR registry key and if it is not already set to 3, it updates the value to 3 to unblock USB ports. This feature enhances the usability of the script by allowing USB devices to be used again.